push 2/3 file DAO

+++ src/main/java/com/carmanagement/entity/Vehicle.java
@@ -3,7 +3,7 @@ package com.carmanagement.entity;
 public class Vehicle extends Product {
     private String VIN;
 
-    public Vehicle(String id_product, String name_product, String brand_product, float price_product, String color_product, int prodution_year_product, String VIN) {
+    public Vehicle(String id_product, String name_product, String brand_product, Integer price_product, String color_product, int prodution_year_product, String VIN) {
         super(id_product, name_product, brand_product, price_product, color_product,  prodution_year_product);
         this.VIN = VIN;
     }
